✨create sidenav_items tag
Create a new `sidenav_items` tag that create a list of `NavItem` object representing the link, icons etc on the side menu, that then will be passed to the existing `nav_item` tag to be created.

Event thought the `nav_items.html` is really simple, I first didn’t find a clean way to just render `nav_item` from the `sidenav_items`, then I was thinking it might be bad after all in case we want to custom `nav_items.html` a bit.